refactor: move tabs layout component into documents group
