+++ frontend/src/components/documents/TabsLayout.spec.ts
