+++ frontend/src/components/documents/TabsLayout.vue
