+++ frontend/src/pages/literature/[documentNumber].vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import IcBaselineSubject from "~icons/ic/baseline-subject";
 import IcOutlineInfo from "~icons/ic/outline-info";
-import type { TabView } from "~/components/TabsLayout.vue";
+import type { TabView } from "~/components/documents/TabsLayout.vue";
 import type { TreeItem } from "~/components/TreeView.vue";
 import { useLiteratureSeo } from "~/composables/useLiteratureSeo";
 import { DocumentKind, type Literature } from "~/types/api";
